+++ android/src/main/java/com/easemob/im_flutter_sdk/EMSDKMethod.java
@@ -124,6 +124,7 @@ public class EMSDKMethod {
     static final String loadMsgWithKeywords = "loadMsgWithKeywords";
     static final String loadMsgWithMsgType = "loadMsgWithMsgType";
     static final String loadMsgWithTime = "loadMsgWithTime";
+    static final String messageCount = "messageCount";
 
     // EMMessage method
     static final String getReactionList = "getReactionList";
@@ -232,6 +233,18 @@ public class EMSDKMethod {
 
     static final String updateHMSPushToken = "updateHMSPushToken";
     static final String updateFCMPushToken = "updateFCMPushToken";
+
+
+    static final String reportPushAction = "reportPushAction";
+    static final String setConversationSilentMode = "setConversationSilentMode";
+    static final String removeConversationSilentMode = "removeConversationSilentMode";
+    static final String fetchConversationSilentMode = "fetchConversationSilentMode";
+    static final String setSilentModeForAll = "setSilentModeForAll";
+    static final String fetchSilentModeForAll = "fetchSilentModeForAll";
+    static final String fetchSilentModeForConversations = "fetchSilentModeForConversations";
+    static final String setPreferredNotificationLanguage = "setPreferredNotificationLanguage";
+    static final String fetchPreferredNotificationLanguage = "fetchPreferredNotificationLanguage";
+
 
     /// EMUserInfoManager 
     static final String updateOwnUserInfo = "updateOwnUserInfo";