+++ lib/im_flutter_sdk.dart
@@ -18,7 +18,7 @@ export 'src/models/em_conversation.dart' show EMConversation;
 export 'src/models/em_cursor_result.dart' show EMCursorResult;
 export 'src/models/em_deviceInfo.dart' show EMDeviceInfo;
 export 'src/models/em_error.dart' show EMError;
-export 'src/models/em_group.dart' show EMGroup;
+export 'src/models/em_group.dart';
 export 'src/models/em_translate_language.dart' show EMTranslateLanguage;
 export 'src/models/em_presence.dart' show EMPresence;
 
@@ -34,5 +34,6 @@ export 'src/models/em_message.dart' hide MessageCallBackManager;
 export 'src/models/em_download_callback.dart' show EMDownloadCallback;
 export 'src/models/em_message_reaction.dart';
 export 'src/models/em_chat_thread.dart';
+export 'src/models/chat_silent_mode.dart';
 
 export 'src/em_message_status_callback.dart' show MessageStatusCallBack;