+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## NEXT
 
+## 3.9.4+3
+
+- 修复 安卓端 `loadAllConversations` crash.
+
+## 3.9.4+2
+
+- 修复 `EMClient.getInstance.startCallback()` 执行时安卓偶现崩溃；
+
+## 3.9.4+1
+- 增加ChatSilentMode;
+
 ## 3.9.4
 - 移除过期Api；